Updated average code to cast data as a double and yield a more accurate result.

+++ csharp-prep/Prep4/Program.cs
@@ -25,7 +25,7 @@ class Program
         }
         if(numbers.Count > 0)
         {
-            average = sum/numbers.Count;
+            average = (double)sum/(double)numbers.Count;
         }
         largest = numbers.Max();
         numbers.Sort();
